Updated test case dtype to float32 to re-check

+++ ivy_tests/test_ivy/test_frontends/test_numpy/test_indexing_routines/test_generating_index_arrays.py
@@ -126,7 +126,7 @@ def test_numpy_tril_indices(
 @handle_frontend_test(
     fn_tree="numpy.tril_indices_from",
     dtype_and_values=helpers.dtype_and_values(
-        dtype=["float16"],
+        dtype=["float32"],
         min_dim_size=3,
         max_dim_size=3,
         min_num_dims=2,
